+++ packages/ai/src/api.test.ts
@@ -14,50 +14,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {
-  ImagenModelParams,
-  ModelParams,
-  AIErrorCode,
-  InferenceMode
-} from './types';
+import { ModelParams, AIErrorCode, InferenceMode } from './types';
 import { AIError } from './errors';
 import {
   getAI,
-  ImagenModel,
   LiveGenerativeModel,
   getGenerativeModel,
-  getImagenModel,
   getLiveGenerativeModel,
   getTemplateGenerativeModel,
-  TemplateGenerativeModel,
-  getTemplateImagenModel,
-  TemplateImagenModel
+  TemplateGenerativeModel
 } from './api';
 import { expect, use } from 'chai';
 import { stub } from 'sinon';
 import { AI } from './public-types';
 import { GenerativeModel } from './models/generative-model';
-import { GoogleAIBackend, VertexAIBackend } from './backend';
-import { getFullApp } from '../test-utils/get-fake-firebase-services';
+import { GoogleAIBackend, AgentPlatformBackend } from './backend';
+import { fakeAI, getFullApp } from '../test-utils/get-fake-firebase-services';
 import { AI_TYPE } from './constants';
 import { logger } from './logger';
 import sinonChai from 'sinon-chai';
 
 use(sinonChai);
-
-const fakeAI: AI = {
-  app: {
-    name: 'DEFAULT',
-    automaticDataCollectionEnabled: true,
-    options: {
-      apiKey: 'key',
-      projectId: 'my-project',
-      appId: 'my-appid'
-    }
-  },
-  backend: new VertexAIBackend('us-central1'),
-  location: 'us-central1'
-};
 
 describe('Top level API', () => {
   describe('getAI()', () => {
@@ -72,25 +49,25 @@ describe('Top level API', () => {
     });
     it('works with options: backend specified, limited use token', () => {
       const ai = getAI(getFullApp(), {
-        backend: new VertexAIBackend('us-central1'),
+        backend: new AgentPlatformBackend('global'),
         useLimitedUseAppCheckTokens: true
       });
-      expect(ai.backend).to.be.instanceOf(VertexAIBackend);
+      expect(ai.backend).to.be.instanceOf(AgentPlatformBackend);
       expect(ai.options?.useLimitedUseAppCheckTokens).to.be.true;
     });
     it('works with options: appCheck option is falsy', () => {
       const ai = getAI(getFullApp(), {
-        backend: new VertexAIBackend('us-central1'),
+        backend: new AgentPlatformBackend('global'),
         useLimitedUseAppCheckTokens: undefined
       });
-      expect(ai.backend).to.be.instanceOf(VertexAIBackend);
+      expect(ai.backend).to.be.instanceOf(AgentPlatformBackend);
       expect(ai.options?.useLimitedUseAppCheckTokens).to.be.false;
     });
     it('works with options: backend specified only', () => {
       const ai = getAI(getFullApp(), {
-        backend: new VertexAIBackend('us-central1')
+        backend: new AgentPlatformBackend('global')
       });
-      expect(ai.backend).to.be.instanceOf(VertexAIBackend);
+      expect(ai.backend).to.be.instanceOf(AgentPlatformBackend);
       expect(ai.options?.useLimitedUseAppCheckTokens).to.be.false;
     });
   });
@@ -169,70 +146,6 @@ describe('Top level API', () => {
     expect(warnStub).to.be.calledWithMatch('generationConfig');
     warnStub.restore();
   });
-  it('getImagenModel throws if no model is provided', () => {
-    try {
-      getImagenModel(fakeAI, {} as ImagenModelParams);
-    } catch (e) {
-      expect((e as AIError).code).includes(AIErrorCode.NO_MODEL);
-      expect((e as AIError).message).includes(
-        `AI: Must provide a model name. Example: ` +
-          `getImagenModel({ model: 'my-model-name' }) (${AI_TYPE}/${AIErrorCode.NO_MODEL})`
-      );
-    }
-  });
-  it('getImagenModel throws if no apiKey is provided', () => {
-    const fakeVertexNoApiKey = {
-      ...fakeAI,
-      app: { options: { projectId: 'my-project', appId: 'my-appid' } }
-    } as AI;
-    try {
-      getImagenModel(fakeVertexNoApiKey, { model: 'my-model' });
-    } catch (e) {
-      expect((e as AIError).code).includes(AIErrorCode.NO_API_KEY);
-      expect((e as AIError).message).equals(
-        `AI: The "apiKey" field is empty in the local ` +
-          `Firebase config. Firebase AI requires this field to` +
-          ` contain a valid API key. (${AI_TYPE}/${AIErrorCode.NO_API_KEY})`
-      );
-    }
-  });
-  it('getImagenModel throws if no projectId is provided', () => {
-    const fakeVertexNoProject = {
-      ...fakeAI,
-      app: { options: { apiKey: 'my-key', appId: 'my-appid' } }
-    } as AI;
-    try {
-      getImagenModel(fakeVertexNoProject, { model: 'my-model' });
-    } catch (e) {
-      expect((e as AIError).code).includes(AIErrorCode.NO_PROJECT_ID);
-      expect((e as AIError).message).equals(
-        `AI: The "projectId" field is empty in the local` +
-          ` Firebase config. Firebase AI requires this field ` +
-          `to contain a valid project ID. (${AI_TYPE}/${AIErrorCode.NO_PROJECT_ID})`
-      );
-    }
-  });
-  it('getImagenModel throws if no appId is provided', () => {
-    const fakeVertexNoProject = {
-      ...fakeAI,
-      app: { options: { apiKey: 'my-key', projectId: 'my-project' } }
-    } as AI;
-    try {
-      getImagenModel(fakeVertexNoProject, { model: 'my-model' });
-    } catch (e) {
-      expect((e as AIError).code).includes(AIErrorCode.NO_APP_ID);
-      expect((e as AIError).message).equals(
-        `AI: The "appId" field is empty in the local` +
-          ` Firebase config. Firebase AI requires this field ` +
-          `to contain a valid app ID. (${AI_TYPE}/${AIErrorCode.NO_APP_ID})`
-      );
-    }
-  });
-  it('getImagenModel gets an ImagenModel', () => {
-    const genModel = getImagenModel(fakeAI, { model: 'my-model' });
-    expect(genModel).to.be.an.instanceOf(ImagenModel);
-    expect(genModel.model).to.equal('publishers/google/models/my-model');
-  });
 
   it('getLiveGenerativeModel throws if no apiKey is provided', () => {
     const fakeVertexNoApiKey = {
@@ -294,11 +207,6 @@ describe('Top level API', () => {
   it('getTemplateGenerativeModel gets a TemplateGenerativeModel', () => {
     expect(getTemplateGenerativeModel(fakeAI)).to.be.an.instanceOf(
       TemplateGenerativeModel
-    );
-  });
-  it('getImagenModel gets a TemplateImagenModel', () => {
-    expect(getTemplateImagenModel(fakeAI)).to.be.an.instanceOf(
-      TemplateImagenModel
     );
   });
 });