+++ packages/auth/demo/rollup.config.js
@@ -19,7 +19,7 @@ import strip from '@rollup/plugin-strip';
 import typescriptPlugin from 'rollup-plugin-typescript2';
 import typescript from 'typescript';
 
-import pkg from './package.json';
+import pkg from './package.json' with { type: 'json' };
 
 const workerPlugins = [
   typescriptPlugin({