+++ packages/firestore/src/core/expressions.ts
@@ -439,8 +439,7 @@ export class CoreListOfExprs implements EvaluableExpr {
 
 function asDouble(
   protoNumber:
-    | { doubleValue: number | string }
-    | { integerValue: number | string }
+    { doubleValue: number | string } | { integerValue: number | string }
 ): number {
   if (isDoubleValue(protoNumber)) {
     return Number(protoNumber.doubleValue);
@@ -1752,7 +1751,7 @@ export class CoreArrayContainsAny implements EvaluableExpr {
             break;
           }
           default:
-            fail(0xae40, { value, search });
+            fail(0xebf3, { value, search });
             break;
         }
       }
@@ -2158,7 +2157,7 @@ export class CoreRegexContains extends StringSearchFunctionBase {
     try {
       const regex = RE2JS.compile(search);
       return EvaluateResult.newValue({
-        booleanValue: regex.matcher(value).find()
+        booleanValue: regex.test(value)
       });
     } catch (RE2JSError) {
       logWarn(
@@ -2817,12 +2816,7 @@ export class CoreTimestampToUnixSeconds extends TimestampToUnix {
 }
 
 type TimeUnit =
-  | 'microsecond'
-  | 'millisecond'
-  | 'second'
-  | 'minute'
-  | 'hour'
-  | 'day';
+  'microsecond' | 'millisecond' | 'second' | 'minute' | 'hour' | 'day';
 function asTimeUnit(unit?: string): TimeUnit | undefined {
   switch (unit) {
     case 'microsecond':