+++ packages/app-check/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @firebase/app-check
 
+## 0.13.0
+
+### Minor Changes
+
+- [`177f08d`](https://github.com/firebase/firebase-js-sdk/commit/177f08d487056e3c88b33071b5162b9a5025241c) [#10210](https://github.com/firebase/firebase-js-sdk/pull/10210) - Remove ability to call `initializeAppCheck` without a `provider`.
+
+### Patch Changes
+
+- Updated dependencies [[`736fa3f`](https://github.com/firebase/firebase-js-sdk/commit/736fa3fe798d6eafc9fe086d6bfbe9cfb49b8994)]:
+  - @firebase/util@1.15.2
+  - @firebase/component@0.7.4
+
 ## 0.12.0
 
 ### Minor Changes